Refactor towards the new world of MutationLog operation.
In this episode, we are removing PathValue as an exposed object, and removing the need of it's set value(), expectValue, and lastObservedValue (only used in element_bindings).

Element bindings now manages it's own need of these functions.

The need for this is so that Model can expose a single "consistency context" for all of MDV such that no one part expect a different value of an object from another -- unless that part manages the different itself (in this case, it can arise if a bindings sets a value)

A "consistency context" is a bad name for a set of guarentees about the semantics of callbacks. The idea is that a group of observers can share a single "consistency context". Notifications are fired for everyone in the group, and no consistency checking is doing between firing of callbacks. The idea here is that a group of observers may know that they can safely listen to observations without danger that one will change a value in a way that will create a bug in another.

BUG=
TEST=

Review URL: http://codereview.chromium.org/7149023

git-svn-id: https://mdv.googlecode.com/svn/trunk@76 e5214b95-32ac-14ea-b9af-4214adf2a604