bump ChangeSummary and allow Observers to call unobserved
This patch goes along with https://github.com/rafaelw/ChangeSummary/commit/75c4e837350228bceec5407fd12830c480ff8359.

The idea here is that chained observation needs a way to clean itself up. These two changes make this a concern of the observers rather than the bindings inside MDV.

It is now the case that for all observed objects, if they have a .unobserved property which is a function, it is invoked when observation ends.

R=arv
BUG=

Review URL: https://codereview.appspot.com/10924044