dont overload binding apis for template binding
john points out that it's weird for TemplateBinding.node to point to the template iterator.

We don't really have a clear delineation of public vs private API right now, but I agree that the iterator reference should really be private.

I'm not prepared yet to go through and start marking private members with trailing underscores. We had that for a while and it just go really ugly.

Hopefully, once we get all of our semantics nailed down, we can write the spec and the IDL and that will make it clear.

r=arv
BUG=

Review URL: https://codereview.appspot.com/12296043