Produce sub-templates synchronously.
This patch make handling of bindings for template bind/if/repeat special, in that they are processed privately as a group (and are not available as public binding names).

Also, the processing order for binding recursive is now post-order, this means that all children elements will be bound and templates produced before a given node binds.

R=arv
BUG=

Review URL: https://codereview.appspot.com/45890043